This PR adds support for key aliases. Still very WIP as I understand more about how this lib works. You're way ahead of me on how you're thinking about advanced caching issues, so I'm trying to find something that fits. 

I will keep this updated with the latest attempt and change it from draft when I think there is something stable to review.

```go
type TestCacheItem struct {
	sturdyc.ISturdyCItem
	key     string
	aliases []string
}

func NewTestCacheItem(key string, aliases []string) *TestCacheItem {
	return &TestCacheItem{
		key:     key,
		aliases: aliases,
	}
}

func (t *TestCacheItem) GetCacheKey() string {
	return t.key
}

func (t *TestCacheItem) GetCacheAliasKeys() []string {
	return t.aliases
}

client.Set("my_key", NewTestCacheItem("value", []string{"alias1", "alias2"}))
client.Get("my_key")
client.Get("alias1")
client.Get("alias2")
```

### Notes
* Item aliases are tracked by the shard that owns the item
* You could consider soft-deprecating `GetManyKeyFn` and `SetManyKeyFn` since they can be covered by implementing the interface.

### Benchmarks

#### before

```
goos: darwin
goarch: arm64
pkg: github.com/viccon/sturdyc
cpu: Apple M1 Max
BenchmarkGetConcurrent-10        7470562               156.6 ns/op               1.000 hits/op
BenchmarkSetConcurrent-10        7944378               180.2 ns/op               0 evictions/op
```

#### after

*These benchmarks don't test any alias load, just the empty logic*

```
goos: darwin
goarch: arm64
pkg: github.com/viccon/sturdyc
cpu: Apple M1 Max
BenchmarkGetConcurrent-10        7158730               162.0 ns/op               1.000 hits/op
BenchmarkSetConcurrent-10        5167750               281.2 ns/op               0 evictions/op
```