@benallfree 2024-12-29
Ok, take a look now. I think I succeeded in keeping things API compatible and performant. This could be good!