@viccon 2024-12-30
Hi, do key aliases mean that you want the ability to store the same value under multiple keys? If not, could you help me understand what a key alias is? I'm not familiar with the term