@benallfree 2024-12-30
Yes, exactly. I need to be able to look up the same cache item by different unique keys. Adding the same item under multiple keys is not ideal because it takes multiple cache entries and some of the keys may change when the item is updated. I would need a way to update them all atomically in a transaction.