@benallfree 2024-12-30
Well this is where perhaps I’m misunderstanding the intent or use case for your lib. I need to be able to retrieve the item by any key without knowing the others. 

Here’s the concrete situation: I run pockethost.io. When a request comes in, it’s either <id>.pockethost.io, <subdomain>.pockethost.io, or <cname>. 

When the request arrives, I make a db query to resolve the requested instance and then proceed accordingly. 

On a cache miss, it should call the API which will return an instance record containing both <id> and <subdomain>, and optionally a <cname> too. From there, I’d like to cache the item so it can be looked up by any of those keys. 

I call it “aliases” because <id> is the key used for caching. Each item takes up just one cache slot, which is important because I wouldn’t want duplicates in the cache.