@ben-manes 2025-01-02
fwiw, this [example](https://github.com/ben-manes/caffeine/tree/master/examples/indexable) of how this feature can be implemented (in Java) which might be helpful. There’s complexity as either a first or second class feature, so the trade offs aren’t clear enough to justify the limitations it would impose. For example refresh would need to update the derived keys, but it’s incompatible with a weak key reference cache. It’s left as an example for now, as I used that approach once 15 years ago making it useful but too rare for deeper consideration at this time.