@viccon 2025-01-02
> fwiw, this [example](https://github.com/ben-manes/caffeine/tree/master/examples/indexable) of how this feature can be implemented (in Java) which might be helpful. There’s complexity as either a first or second class feature, so the trade offs aren’t clear enough to justify the limitations it would impose. For example refresh would need to update the derived keys, but it’s incompatible with a weak key reference cache. It’s left as an example for now, as I used that approach once 15 years ago making it useful but too rare for deeper consideration at this time.

Thanks! Yes, I've been thinking about this a bit, and there would also be some complexity involved when you're using this in combination with a distributed storage like Redis. The cache currently has the `WithDistributedStorage` and `WithDistributedStorageEarlyRefreshes` options that would need to be taken into account as well. The refresh coalescing would also need additional work because I would only want to add one of the aliases during the buffering, and then update all of them when I get the response. Right now, I'm leaning toward not adding this functionality, because the cache is already quite complex, and every additional feature must work with all the other options it currently supports.