@benallfree 2025-01-02
Thanks for considering it! I agree it's an usual case that would slow things down and potentially break some features.

Closing this PR until we have a better approach.