@ben-manes 2025-01-02
> Closing this PR until we have a better approach.

I think you could layer it above the cache as a more limited, specialized one just for your needs. That’s what the example shows so that the underlying library doesn’t maintain the feature while unblocking any users who need it without deep integration features. That’s what I meant as second class where it’s possible to, just left to users for implementing and maintaining.